+++ src/App.tsx
@@ -4,7 +4,6 @@ import { Bot, FolderCode } from "lucide-react";
 import { api, type Project, type Session } from "@/lib/api";
 import { TabProvider, useTabContext } from "@/contexts/TabContext";
 import { AccountsProvider } from "@/contexts/AccountsContext";
-import { AppCapabilitiesProvider } from "@/contexts/AppCapabilitiesContext";
 import { ThemeProvider } from "@/contexts/ThemeContext";
 import { AppFontProvider } from "@/contexts/AppFontContext";
 import { MessageRenderingProvider } from "@/contexts/MessageRenderingContext";
@@ -592,14 +591,12 @@ function App() {
     <AppFontProvider>
       <ThemeProvider>
         <MessageRenderingProvider>
-          <AppCapabilitiesProvider>
-            <AccountsProvider>
-              <TabProvider>
-                <AppContent />
-                <StartupIntro visible={showIntro} />
-              </TabProvider>
-            </AccountsProvider>
-          </AppCapabilitiesProvider>
+          <AccountsProvider>
+            <TabProvider>
+              <AppContent />
+              <StartupIntro visible={showIntro} />
+            </TabProvider>
+          </AccountsProvider>
         </MessageRenderingProvider>
       </ThemeProvider>
     </AppFontProvider>